[FE0012] Update Feedback UI
