--- src/pages/DetailProduct/DetailProductScreen.tsx
@@ -1,9 +1,0 @@
-import DetailProductComponent from "./DetailProductComponent";
-
-const DetailProductScreen = () => {
-    return (
-        <DetailProductComponent />
-    );
-}
-
-export default DetailProductScreen;
